Move more collection management related components into the existing collectionManagement subfolder
