+++ app/packs/src/components/collectionManagement/CollectionManagement.js
