+++ app/packs/src/components/collectionManagement/CollectionTree.js
@@ -3,7 +3,7 @@ import { Button, OverlayTrigger, Badge, Glyphicon, Tooltip } from 'react-bootstr
 import update from 'immutability-helper';
 import CollectionStore from 'src/stores/alt/stores/CollectionStore';
 import CollectionActions from 'src/stores/alt/actions/CollectionActions';
-import CollectionSubtree from 'src/components/CollectionSubtree';
+import CollectionSubtree from 'src/components/collectionManagement/CollectionSubtree';
 import UIActions from 'src/stores/alt/actions/UIActions';
 import InboxActions from 'src/stores/alt/actions/InboxActions';
 import LoadingActions from 'src/stores/alt/actions/LoadingActions';
